+++ mono/coder/views.py
@@ -4,10 +4,16 @@ from django.http import Http404
 from django.shortcuts import get_object_or_404
 from django.views.generic.base import TemplateView
 from django.views.generic.detail import DetailView
+from pygments import highlight
 from pygments.formatters import \
     HtmlFormatter  # pylint: disable=no-name-in-module
+from pygments.lexers import get_lexer_by_name
+from rest_framework.generics import RetrieveUpdateAPIView
 
-from .models import LANGUAGE_CHOICES, Configuration, Snippet, Tag
+from .models import (
+    LANGUAGE_CHOICES, STYLE_CHOICES, Configuration, Snippet, Tag,
+)
+from .serializers import ConfigurationSerializer
 
 
 class RootView(TemplateView):
@@ -27,11 +33,28 @@ class SnippetListView(LoginRequiredMixin, TemplateView):
         return Snippet.objects.filter(created_by=self.request.user)
 
     def get_context_data(self, *args, **kwargs):
+        code = [
+            'def fib(n):',
+            '   """Print Fibonacci sequence"""',
+            '    a, b = 0, 1',
+            '    while a < n:',
+            '        print(a, end=\' \')',
+            '        a, b = b, a + b',
+            'fib(1000)',
+        ]
         context = super().get_context_data(*args, **kwargs)
         context['languages'] = LANGUAGE_CHOICES
+        context['styles'] = STYLE_CHOICES
         context['colors'] = Tag.Color.choices
-        config: Configuration = Configuration.objects.get_or_create(user=self.request.user)[0]
-        context['snippet_css'] = HtmlFormatter(style=config.style).get_style_defs('.highlight')
+        context['all_styles_css'] = ''.join(
+            HtmlFormatter(style=style[0]).get_style_defs(f'.{style[0]}')
+            for style in STYLE_CHOICES
+        )
+        context['demo_code'] = highlight(
+            '\n'.join(code),
+            lexer=get_lexer_by_name('python', stripall=True),
+            formatter=HtmlFormatter(linenos=True),
+        )
         return context
 
 
@@ -57,3 +80,12 @@ class SnippetPublicView(TemplateView):
         context['snippet'] = snippet
         context['snippet_css'] = HtmlFormatter(style='monokai').get_style_defs('.highlight')
         return context
+
+
+class ConfigAPIView(RetrieveUpdateAPIView):
+    """View to read or update user config"""
+
+    serializer_class = ConfigurationSerializer
+
+    def get_object(self):
+        return Configuration.objects.get_or_create(created_by=self.request.user)[0]